`taxstatus` table added, relevant files modified.
